fix: split route components for React Doctor
