+++ src/routes/$lang.tsx
@@ -4,24 +4,16 @@ import {
 	Outlet,
 	redirect,
 } from "@tanstack/react-router";
-import { bind as bindCuelume } from "cuelume";
-import { type ReactNode, useEffect } from "react";
-import AppUpdateToast from "../components/AppUpdateToast";
-import Footer from "../components/Footer";
-import Header from "../components/Header";
-import NotFoundPage from "../components/NotFoundPage";
-import StarPopup from "../components/StarPopup";
+import { useEffect } from "react";
+import AppChrome from "../components/AppChrome";
+import LangNotFound from "../components/LangNotFound";
 import { I18nProvider } from "../i18n/context";
 import { isLang } from "../i18n/context-value";
 import { detectPreferredLang } from "../i18n/detect-lang";
-import { useLang, useT } from "../i18n/hooks";
-import { initializeSound } from "../lib/sound";
 import { ssrDuringBuildPrerender } from "../lib/ssr";
-import { getDistinctId, identify, setSessionData } from "../lib/umami";
 import { createNoIndexHead } from "../seo/head";
 import { buildNotFoundMeta } from "../seo/meta";
 import { ThemeProvider } from "../theme/context";
-import { useTheme } from "../theme/hooks";
 
 export const Route = createFileRoute("/$lang")({
 	ssr: ssrDuringBuildPrerender,
@@ -96,68 +88,5 @@ function LangLayout() {
 				</AppChrome>
 			</I18nProvider>
 		</ThemeProvider>
-	);
-}
-
-function LangNotFound() {
-	const { lang: rawLang } = langRouteApi.useParams();
-	const lang = isLang(rawLang) ? rawLang : "es";
-
-	useEffect(() => {
-		document.documentElement.lang = lang;
-	}, [lang]);
-
-	return (
-		<ThemeProvider>
-			<I18nProvider initialLang={lang}>
-				<AppChrome>
-					<NotFoundPage />
-				</AppChrome>
-			</I18nProvider>
-		</ThemeProvider>
-	);
-}
-
-function SessionTracker() {
-	const { lang } = useLang();
-	const { theme } = useTheme();
-
-	useEffect(() => {
-		const id = getDistinctId();
-		if (id) identify({ id });
-	}, []);
-
-	useEffect(() => {
-		setSessionData({ lang, theme });
-	}, [lang, theme]);
-
-	return null;
-}
-
-function AppChrome({ children }: { children: ReactNode }) {
-	const t = useT();
-
-	useEffect(() => {
-		bindCuelume();
-		initializeSound();
-	}, []);
-
-	return (
-		<div className="bg-surface text-fg flex min-h-screen min-h-svh min-h-dvh flex-col font-sans">
-			<SessionTracker />
-			<a
-				href="#main-content"
-				className="focus-visible:ring-accent sr-only focus:not-sr-only focus:fixed focus:top-2 focus:left-2 focus:z-[100] focus:rounded-md focus:bg-accent focus:px-4 focus:py-2 focus:font-semibold focus:text-accent-fg focus-visible:ring-2 focus-visible:outline-none"
-			>
-				{t.header.skipToContent}
-			</a>
-			<Header />
-			<main id="main-content" tabIndex={-1} className="flex-grow">
-				{children}
-			</main>
-			<Footer />
-			<AppUpdateToast />
-			<StarPopup />
-		</div>
 	);
 }
